Add a distance named argument to read_token.
This can be used to specify that the next token must be within a
certain distance, either in lines or columns, of the last token. So,
if you want to ensure that two tokens don't have space between them
(if you want to make a difference between spaced and unspaced) then
you can specify a distance and read the token using that.