Add argv[N] / $N, [bits], and named registers as arguments.
Named registers: ip, ebp, esp, rp.

argv[N] and $N both refer to argument registers. It's just the first
argument register + N. So, $0 is the first argument, $5 the sixth, and
so on.

[bits] arguments are bitset arguments, essentially. So, for example,
[0] is the same as 0x1, [1] is the same as 0x2, [3] 0x4, and so on.
This is most useful for multiple bits, however. So, [0 1 2 3] is 0xF.
This is mainly intended to aid when passing register masks to CALL,
PUSH, and POP instructions, since having to pass bin/hex/etc. literals
isn't very pleasant. Note that you can refer to specific registers in
bitset arguments, like [rp argv[0] argv[4] $20] and that's perfectly
fine, so there's no need to keep track of what argv[N] or rp or what
have you map to, though it's probably useful (but it does make it
easier to reassemble bytecode after breaking changes are made to the
VM and such, since you're no longer referring to registers by their
number in the VM, though they will be converted to their VM number).