Add new data block types and fields.
_data is simply a buffer containing all data loaded. It can be until
the unit is in use by the VM. This isn't really an issue since the VM
will copy the entire unit over (because it may need to relocate data).
I'll probably tweak this later to ensure data blocks are 8-byte aligned
or something along those lines (since it's the max peek-able size).

_data_blocks is the table of all data block IDs, offsets (into _data),
and sizes.

_data_relocations holds all data relocations. Might be empty. Used by
the VM (vm_state_t) when it copies a unit to relocate all block IDs to
whatever the VM assigns to them, since it's not required to preserve
the IDs and it's easy enough to reload everything.