Add esp preservation.
This probably looks a little funny since the preserved esp is ebp
after updating it to account for arguments on the stack. Basically,
after the call, esp will be set to the stack's original size sans
arguments, so neither the caller nor the callee are required to clean
up the stack.