Add primitive let[*!] blocks.
Syntax is fairly simple:
  let name_a, name_b, name_c[, ...] {

  }

Inside a let block, any defined name is assigned to general purpose
registers. Currently, this just means they're assigned in order.

Default let block behavior is that if a volatile register is aliased,
it is pushed at the beginning of the let block and popped at the end to
avoid stomping variables above it in the stack.

Because the default behavior isn't always desirable, there are two
optional forms of let blocks, depending on the symbolic suffix of the
keyword:

  let* -- preserve no registers. This simply maps the names as aliases
    of registers and nothing else.
  let! -- preserve all registers by pushing them onto the stack and
    restoring their values when done.

There is currently no way to assign names to specific registers, though
I'll probably work on that next (since registers are in memory, you may
at some point decide that it makes sense to try to have sequential
access to certain registers).

ip, ebp, esp, and rp are all reserved names and cannot be re-defined by
a let block.