Add explicit register aliasing.
Any let block or a function's arguments may use the syntax varname(reg)
to alias a specific register. Automatic Register allocation will skip
already-aliased registers, as usual.

E.g.,

  function name(name_1(%24), name_2) { ... }
  let name_1(%24), name_2 { ... }

Where name_1 is aliased to register %24 and name_2 is aliased to
register %4. Reserved registers may not be aliased and named registers
may not be used as register targets. As such, you cannot do a(b) where
b is any named register, nor can you do a(%3) because %3 is rp and a
reserved register.