Refactor static memory allocation/loading.
Getting blocks now done through get_block_info. get_block is a wrapper
around it with a permissions check.

Allocating a block is now done using realloc_with_flags, allowing the
caller to provide the flags for that block (as in prepare_unit when
loading static memory blocks as individual allocations). This was done
specifically to avoid a write-after-free issue in the unit destructor
where unit memory was still in use when it shouldn't have been.

get_block_info optionally returns a block, so its result should be
checked. If get_block_info returns nothing in get_block, then get_block
now returns nullptr instead of raising an exception for attempting to
access a block that wasn't in _blocks using map.at(K).