Implement DEFER instruction.
Right now, the loss of the call stack is accidental -- the VM will not
currently recreate a stack or permit partial execution of something
(i.e., only executing N instructions before yielding to the host). I'll
be fixing this later, but for now this is expected behavior and
accidentally benefits this feature. Once it's fixed, this will need to
have its call stack explicitly flattened.