run until complete in call_function_nt().
TRAP may need the ability to indicate a specific flag. As is, it now
behaves as a sort of yield, returning control to whatever called run().
As such, trap should be used, for now, as a way of indicating that the
VM is stopping itself to give the host process a chance to do stuff.

Currently, the JOIN instruction ignores the result of run() since it's
not intended to handle yields from a joined thread. call_function_nt is
an oddity since it's a wrapper around other VM behavior that allows
simple calls into the VM, but also doesn't do anything with a TRAP
other than to ignore and rerun the VM.