Perform internal symbol lookup by hash rather than string.
The chance of a conflict arising grows, but thus far snow-hash hasn't
resulted in conflicts with the sort of data it's made for (identifiers)
so I'm inclined to continue using it.

It's probably feasible to also switch to unordered_map, but I haven't
spent much time looking at that yet. I'd assume there's plenty of cases
where I can just replace a map with a vector and either iterate or do
a binary search, too.