Fix NEG/NOT.
NOT: confusing because of the ~ (highlighter wanted to treat this as
a dtor and it's kind of unclear what's going on at first glance, so
I'm just wrapping it in parens).

NEG: Always casted to fp64, which is not the desired behavior now that
it's possible to have integral types.