Adjust rounding behavior.
For all rounding instructions, the result is always, necessarily, a
float. If the input is not a double, it's casted to a double and
returned.

Not yet sure if the casted result should also be rounded, since I
suspect there'll be precision issues for integers that exceed the
precision of a double.

Alternatively, the result of rounding an integer could also be an
undefined value, but that sounds weird. Maybe just cast to signed
integer instead?